% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasl-package.r, R/gasl.R
\docType{package}
\name{gasl}
\alias{gasl}
\alias{gasl-package}
\alias{gasl}
\title{gasl.}
\usage{
gasl(X, Y, factor_vars, factor_args = "lm", covariate_args = "lm",
  gam_control = gam::gam.control(trace = F, bf.epsilon = 0.001))
}
\arguments{
\item{X}{predictor variables as a matrix}

\item{Y}{outcome variable, a vector.}

\item{factor_vars}{a character vector indicating which variables to include in the factor fit. All other variables in \code{X} will be included in the covariate fit.}

\item{factor_args}{Either \code{'lm'} (for a linear smoother) or a list of arguments for \code{\link{origami::origami_SuperLearner}}, indicating the smoother used for the factor variables}

\item{covariate_args}{Either \code{'lm'} (for a linear smoother) or a list of arguments for \code{\link{origami::origami_SuperLearner}}, indicating the smoother used for the covariate variables}

\item{gam_control}{A list of parameters to control the gam fit. Usually generated with \code{\link{gam::gam.control}}}
}
\value{
A gasl object
}
\description{
gasl.

GAM with Super Learner smoother
}
